Improvements to goal tracking and weight logging views
